Fix bug in review_section
/2014/reviews/section/talks/ and similar pages were not showing all
the proposals they should have, due to a bug in the review_section
view.

Also changed the review_assignments view to grant access to all
superusers, for consistency with other views.

Also added some doc about how to enable proposal submission, and
added a couple help text strings for the date fields that control
the dates submissions are open, and fixed a bunch of PEP-8
problems.