Bad rendering on proposal details if no tags
If a proposal has no tags, on the Proposal Details page, the
value of the Category gets displayed shifted up so it looks like
that's the tags value.

Apparently this is a problem with all the "dd" tags on this page,
because I see all the other ones have &nbsp; appended.

So, doing the same for tags.