Require travel plans on financial aid form
More specifically, require travel plans on the financial aid
form if and only if a travel grant is requested. Implemented
by leaving the form field technically not required, adding
form validation that looks at whether a travel grant is requested,
and marking the travel plans field with 'required' styling
manually.