385 Financial aid state machine
Depending on the current state of someone's application,
provide appropriate buttons on the dashboard for the user
to move their application to any other state that they
are allowed to move it to, given its current state.

E.g. if application is in submitted state, they can
withdraw it. If offered, they can accept or decline.
Etc.

Fixes #385