Fix fab dbshell
'fab server manage_run:dbshell' was not working. I guess there
wasn't enough information in the Django settings for Django to
invoke psql correctly to connect to our databases the way they
are set up in the PSF infrastructure.

Added a new fab task 'dbshell' that extracts the needed info
the same way we do before running a dump, and invokes psql
directly.  Also tweaked 'manage_run' so if 'dbshell' is passed
as the command, it uses the new task.