Move session cache to a different Redis DB
Right now, the general cache and the session cache are in the
same DB, so if we want to clear cached pages, it results in everyone's
sessions (and logins) being deleted.

By putting the sessions in a different Redis DB, we should
be able to clear the general cache without affecting sessions.