547 Fix error when changing proposal status
For edusummit proposals especially.

The fix is really the 3 lines changed to 1
in reviews/views.py, but adding a test for this
problem turned up some inconsistencies in other
places that I went ahead and addressed.

I also fixed any PEP-8 problems in the files that
I was changing anyway.

Fixes #547