551 Fix error submitting reviews
Added test to recreate problem.

I think this was something that broke when Django was
upgraded, but without a unit test, we weren't aware of
it.

Change from passing a class instance to .update() to
using the (documented) F function to do the same thing
that seemed to be intended.

Also work around Django's annoying behavior of raising
an exception when a reverse relation has no object.

Fixes #551