Allow speaker photo uploads with non-ASCII filenames
Just make up a random, ASCII filename for each one.

Without this change, I could reproduce the problem with
the locale set to C and not with the locale set to
en_US.UTF-8, so running with locale en_US.UTF-8 ought
to solve it, but on the actual servers it doesn't seem
to. Try this too.