Add auto-loading of fixtures at provision time
In order to get this working, a number of fixtures that have been broken
for some time needed fixing, and a few needed to be deleted altogether
because they held data that is now populated during the initial Django
"migrate" step.

This also makes migration and fixture-loading into one-time steps that
only happen right after the database is created during the initial
provision.  To give developers a stable environment once a VM exists, it
seems useful for these steps to not be run over again automatically,
even if "provision" gets run again.