Fix #580 by turning off Biblion markup entirely
Having a markup system was, in the end, annoying: I always write blog
posts in another system anyway, and then just want to cut and paste the
raw HTML over in to the PyCon site.  But instead I was having to rewrite
it by hand in Markdown so that I could submit it.  Now, I can copy the
raw HTML and paste it directly into the Django Admin form in a single
step.  Things will be simpler, and far more robust.  And we get to say
"goodbye" to the literal <h2> tags that the broken Symposion markup
processor was writing into the blog posts!