Per committee chair, make vote comments optional
This is both a semantic shift that the committee chair would like,
because they are not sure that it is really useful to make every single
busy reviewer stop to write a comment on every single talk they vote on
if they feel the topic has already been well enough covered by the
existing committee comments, and also a serious UI fix since, if the
reviewer fails to comment, they cannot see the form error when the page
reloads — it will be hidden by the fact that the voting form is on the
second tab of the page, with no notification that an error is sitting
there that needs to be read.  This fixes #588.