Fix user selection for speaker add
When adding a speaker in the admin, the staff person had to
pick a user from a huge dropdown with all the users, unsorted.

Change 'user' to a raw id field, meaning to pick a user, the staff member clicks a magnifying glass icon next to the field and gets a popup listing all the users in an admin list page with sortable columns and search.