Bugfix: finally make Markdown rendering consistent
Previously, we were delivering Markdown to the page un-quoted, as though
it were real HTML, and then calling `.html()` to turn it back into text.
This round-trip through the browser's HTML parser often damaged the
Markdown, as many Markdown conventions — like the use of `>` to create a
blockquote — are invalid HTML.  This often meant that Markdown would
look fine during editing, but look broken and terrible once the person
hit "Save".  So we now correctly deliver the markdown fully quoted to
protect every character from the browser, and call `.text()` to receive
the content for rendering.