Update importer for 2017
Since we started fresh this year with a new database, instead of simply
updating 2016's, a few more tables had to be populated with data, and
the extra columns in the imported CSV data had to be added slightly
later once the data they needed was present.

Also, adjusts the break times for the slightly different 2017 schedule.