Fix import so it creates session buttons
The "badge" buttons in the schedule need to correctly be linked to the
right `schedule_id`. My 2016 import script was based on the old
pre-populated database where, at wit's end, I could just manually
reference the ID of the session I knew I wanted the slots associated
with. But in 2017, where we started with a fresh database, I actually
have to join across 2 more tables to get the correct id. It is nice to
pivot away from a hard-wired value to a real query. But, wow, this site
seems to have way too many tables.