Fix program export 500 error: remove old fields
Since we have simplified the application form this year and all of the
old "audience level" and "category" fields are NULL, the exporter dies
trying to access their related objects. So, as a simplest-possible fix,
this removes them from the output — a change that we should be able to
quickly revert in a future year if we want them back.