Pull talk descriptions, authors from proposal
When staff edit a proposal — for example, on behalf of a presenter who
notices a typo, or a sponsor who wants to update the description of
their Sponsor Workshop — they have to ask for an import to be run so
that the title, speakers, and description fields are copied over from
the proposal table to the presentation table.

Why make them do that?  Let's just display the details from the proposal
to begin with, and consider that the table-of-record for this data.